+++ src/static/helpers/slasHelper.ts
@@ -425,7 +425,7 @@ export async function loginGuestUser(
  * @param credentials.username - the id of the user to login with.
  * @param credentials.password - the password of the user to login with.
  * @param credentials.clientSecret? - secret associated with client ID
- * @param parameters - parameters to pass in the API calls. Custom parameters can be passed on by adding a property on the `parameters` object starting with `c_`, and they will be passed to `authenticateCustomer` call.
+ * @param parameters - parameters to pass in the API calls.
  * @param parameters.redirectURI - Per OAuth standard, a valid app route. Must be listed in your SLAS configuration. On server, this will not be actually called. On browser, this will be called, but ignored.
  * @param parameters.usid? - Unique Shopper Identifier to enable personalization.
  * @param parameters.dnt? - Optional parameter to enable Do Not Track (DNT) for the user.
@@ -449,7 +449,7 @@ export async function loginRegisteredUserB2C(
     redirectURI: string;
     usid?: string;
     dnt?: boolean;
-  } & CustomQueryParameters,
+  },
   options?: {
     body?: CustomRequestBody;
   }
@@ -472,14 +472,12 @@ export async function loginRegisteredUserB2C(
   const authorization = `Basic ${stringToBase64(
     `${credentials.username}:${credentials.password}`
   )}`;
-  const {dnt, usid, redirectURI, ...restOfParams} = parameters;
+  const {dnt, usid, redirectURI} = parameters;
   const opts = {
     headers: {
       Authorization: authorization,
     },
     parameters: {
-      // putting this at the top to avoid any overriding on the required query param below
-      ...restOfParams,
       organizationId: slasClient.clientConfig.parameters.organizationId,
     },
     body: {